compile, test, run successfully. But no data returns from h2

+++ src/main/java/com/example/springboot/repository/PersonRepository.java
@@ -1,6 +1,7 @@
 package com.example.springboot.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+// import org.springframework.data.repository.CrudRepository;
 
-public interface PersonRepository extends JpaRepository<Person, Integer> {
+public interface PersonRepository extends JpaRepository<Person, Long> {
 }
